move TankSettings struct into structs file
